Slightly improved functionality for AddPlayer
Now the function will not let an empty name. It will exit the game if the player clicks Cancel or X of the modal window. The function does not let two people to have the same name (case sensitive). So a player can be called Rosi and rosi and taht will work