Changes on Restart
Now in case of draw the players get a message, the score is not added, new game starts. The winner music is changed. Try catch and the exceptions were not needed where playSound() is not called. Added a variable playedTurns that resets on reset, on a new game in case of a draw or in case of a winner